Moved court allocation to submission and refresh court data before payments
